Added Prototype Box holder
Including collisions and extra functions in entity. It does appear that this lowers the framerate slighty.

Visibly there is now a box in the top right corner of the starting platform. When the player steps on it it dissapears, when you push a rigid on the empty space it looks to start hanging in thin air. 

What actually happens is the the boxholder box is "turned visible" (Actually returned from other location) and the rigid is put outside of the screen's range after which it respawns.